backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==46.0.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
exceptiongroup==1.3.1
fastapi==0.134.0
frozenlist==1.8.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
h11==0.16.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
multidict==6.7.1
netaddr==0.8.0
netifaces==0.11.0
notify2==0.3
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==9.0.1
pop-transition==1.1.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
ptyprocess==0.7.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==3.0
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
starlette==0.52.1
systemd-python==234
twilio==9.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uritemplate==4.2.0
urllib3==1.26.5
uvicorn==0.41.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.22.0
zipp==1.0.0
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
pytz
python-dateutil